Split ring-based ticketer and kx into their own files
